move test_fill_axis.cc into test_construction.cc
